refactor: split private state sanctuary module
